added some authentications with firebase

+++ src/app/recipes/recipes-detail/recipes-detail.component.ts
@@ -2,6 +2,7 @@ import {Component, OnInit} from '@angular/core';
 import {RecipeService} from "../recipe.service";
 import {Recipe} from "../recipe.model";
 import {ActivatedRoute, Params, Router} from "@angular/router";
+import {AuthService} from "../../auth/auth.service";
 
 @Component({
   selector: 'app-recipes-detail',
@@ -12,7 +13,10 @@ export class RecipesDetailComponent implements OnInit {
   recipe: Recipe;
   id: number;
 
-  constructor(private recipeService: RecipeService, private route: ActivatedRoute, private router: Router) { }
+  constructor(private recipeService: RecipeService,
+              private route: ActivatedRoute,
+              private router: Router,
+              private authService: AuthService) { }
 
   ngOnInit() {
     this.route.params.subscribe(
@@ -24,17 +28,27 @@ export class RecipesDetailComponent implements OnInit {
   }
 
   onAddToShoppingList(){
-    this.recipeService.addIngredientsToShoppingList(this.recipe.ingredients);
+    if (this.isAuthenticated()) {
+      this.recipeService.addIngredientsToShoppingList(
+        this.recipe.ingredients);
+    }
   }
 
   onEditRecipe(){
-    this.router.navigate(['edit'], {relativeTo: this.route})
+    this.router.navigate(['edit'],
+      {relativeTo: this.route})
   }
 
   onDeleteRecipe(){
-    this.recipeService.deleteRecipe(this.id);
-    if (!this.router.navigate(['recipes'])){
-      console.log('Error Navigating to recipes');
+    if (this.isAuthenticated()) {
+      this.recipeService.deleteRecipe(this.id);
+      if (!this.router.navigate(['recipes'])) {
+        console.log('Error Navigating to recipes');
+      }
     }
+  }
+
+  private isAuthenticated() {
+    return this.authService.isAuthenticated();
   }
 }
